jwt_authn: replace with wasm buildable base64
